+++ graceful.go
@@ -27,26 +27,6 @@ func isMaster() bool {
 	return len(os.Getenv(listenEnv)) == 0
 }
 
-type fileFunc interface {
-	File() (f *os.File, err error)
-}
-
-func listenerFiles(listeners []net.Listener) ([]*os.File, error) {
-	files := make([]*os.File, 0, len(listeners))
-	for _, l := range listeners {
-		fd, ok := l.(fileFunc)
-		if !ok {
-			return nil, errors.New("no File() method for " + l.Addr().String())
-		}
-		f, err := fd.File()
-		if err != nil {
-			return nil, err
-		}
-		files = append(files, f)
-	}
-	return files, nil
-}
-
 func restoreListeners(envvar string) ([]net.Listener, error) {
 	nfds, err := strconv.Atoi(os.Getenv(envvar))
 	defer os.Unsetenv(envvar)
