+++ graceful_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package cmd
+
+func isMaster() bool {
+	return true
+}
